factor out new v2 batch ingest types and biz logic into own module
